Split app into server.py and app folder
