+++ console/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+
+app = Flask(__name__)
+app.config['USE_MLSD'] = True
+
+import console.views  # noqa
